+++ mne_connectivity/utils/__init__.py
@@ -1,3 +1,3 @@
 from .docs import fill_doc
 from .utils import (check_indices, degree, seed_target_indices,
-                    parallel_loop)
+                    parallel_loop, _prepare_xarray_mne_data_structures)